move TimestampParam to non-td operator package
