+++ digdag-standards/src/main/java/io/digdag/standards/operator/td/TdPartialDeleteOperatorFactory.java
@@ -8,6 +8,7 @@ import io.digdag.spi.Operator;
 import io.digdag.spi.OperatorFactory;
 import io.digdag.spi.TaskExecutionContext;
 import io.digdag.spi.TaskRequest;
+import io.digdag.standards.operator.TimestampParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
